Add a "refref" facility to internally take VCL references
So it turns out, that - still - we can deadlock raching between VCL
events and director deletions (see
https://github.com/varnishcache/varnish-cache/issues/4140). This commit
was originally written to fix #108, but it now enables us to take a
different route again: Do not pthread_join() lookup threads, but rather
detach them and make sure _they_ return the reference before terminating.

The remainder of the commit message is an edit in light of the new
insights:

Our lookup threads may continue for a bit after a VCL_COLD event has
been received: dom_event() sets dom->status to DYNAMIC_ST_DONE while
the lookup thread may be blocking in the resolver or updating a
domain.

So, lookup threads might try to create backends even after the VCL has
gone cold, in which case VRT_AddDirector() would panic with "Dynamic
Backends can only be added to warm VCLs" - _unless_ a reference is
present on the VCL to make the VCL state change go through
VCL_TEMP_COOLING, in which case VRT_AddDirector() just returns NULL.

vmod_dynamic took a VCL reference since
99912bc1cd4662e4e62f4ed24cd759ad22c63ee5 back in May 2019, but before
the merge of https://github.com/varnishcache/varnish-cache/pull/4037
it did not work correctly, because the VCL reference was returned
before all threads had terminated.

The naive idea to solve this would be to take/release a VCL reference
in each lookup thread (or when starting/stopping it), but references
can only be taken from the CLI context (ASSERT_CLI() in
VRT_VCL_Prevent_Discard). Also, to support use of the .resolve() type
method from vcl_init{}, we needed to depart from the original concept
of running resolver threads only for warm vcls.

We get out of this situation by using a reference on top of the VCL
reference, which we can take even before we have the VCL reference.
The VCL reference is returned once all internal "referef"s are returned.

Ref #108

Probably fixes #117 also